Fix scoping issue on replace function.
